requirements.txt
pillow
ursina
numba
panda3d-gltf
panda3d-simplepbr
